Take care of big values for orders

+++ src/de/matdue/isk/EveApiUpdaterService.java
@@ -262,7 +262,7 @@ public class EveApiUpdaterService extends WakefulIntentService {
 		orderWatch.price = marketOrder.price;
 		orderWatch.volEntered = marketOrder.volEntered;
 		orderWatch.volRemaining = marketOrder.volRemaining;
-		orderWatch.fulfilled = 100 - ((marketOrder.volEntered != 0) ? (100 * marketOrder.volRemaining / marketOrder.volEntered) : 0);
+		orderWatch.fulfilled = (int) (100 - ((marketOrder.volEntered != 0) ? (100 * marketOrder.volRemaining / marketOrder.volEntered) : 0));
 		if (orderWatch.fulfilled == 100 && marketOrder.volRemaining != marketOrder.volEntered) {
 			// Make sure nearly completed orders do not reach 100%
 			// because of rounding
